✨ refactor : SectionTitle 컴포넌트를 common 폴더로 이동하고 관련 경로 수정 #52
