+++ src/components/sections/common/SectionTitle.tsx
